refactor: renamed & added data files
`levels.txt` was renamed to `level_reports.txt` to better reflect the nature of the data contained in the file.

`sample.txt`, containing sample data from the Advent of Code puzzle description to illustrate the instructions, was added as well as a set of test data to validate the algorithm.